backend/requirements.txt
# ==================== 核心框架 ====================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0
python-multipart==0.0.12

# ==================== 数据库 ====================
sqlalchemy==2.0.36
pymysql==1.1.1
psycopg2-binary==2.9.10

# ==================== 缓存 ====================
redis==5.2.0

# ==================== HTTP 客户端 ====================
httpx[socks]==0.28.0

# ==================== HTML 解析 ====================
beautifulsoup4>=4.12.0

# ==================== AI/LLM ====================
openai>=1.55.3

# ==================== 文档处理 ====================
python-docx==1.1.2

# ==================== 认证 & 安全 ====================
PyJWT==2.10.0
passlib==1.7.4

# ==================== 邮件 ====================
Jinja2==3.1.4

# ==================== 支付 ====================
alipay-sdk-python==3.7.1018
rsa>=4.9  # rsa 4.7.2 不兼容 pyasn1 0.6+，4.9+ 已修复
# Paddle payments use REST API via httpx (no additional SDK needed)

# ==================== 配置管理 ====================
python-dotenv==1.0.1

# ==================== 测试工具 ====================
pytest==8.3.0
pytest-asyncio==0.24.0
